refactor: 아이템 Query Key Factory 구조 적용 및 폴더 구분
