+++ src/app/(route)/items/_components/ItemList.tsx
@@ -4,7 +4,7 @@ import React, { useState, useEffect } from 'react'
 import { useRouter } from 'next/navigation'
 import Image from 'next/image'
 import { ItemType } from '@/app/_types/item.type'
-import useItemListData from '@/app/_hook/api/useItemListData'
+import useItemListData from '@/app/_hook/api/items/useItemListData'
 import useGetSearchParam from '@/app/_hook/common/useGetSearchParams'
 import { useInView } from 'react-intersection-observer'
 import SortBox from './SortBox'
