+++ src/app/_hook/api/items/index.ts
@@ -1,0 +1,6 @@
+export const itemKeys = {
+  all: ['items'] as const,
+  itemList: (keyword: string, sortOptions: string) =>
+    [...itemKeys.all, keyword, sortOptions] as const,
+  detail: 'itemDetail' as const,
+}
