+++ src/app/_hook/api/items/useItemDetail.ts
@@ -1,3 +1,5 @@
+import { itemKeys } from '.'
+
 export async function fetchItemDetail(itemId: number) {
   const res = await fetch(
     `${process.env.NEXT_PUBLIC_BASE_URL}/api/items/${itemId}`,
@@ -6,7 +8,7 @@ export async function fetchItemDetail(itemId: number) {
       headers: {
         'Content-Type': 'application/json',
       },
-      next: { tags: ['itemDetail'] },
+      next: { tags: [itemKeys.detail] },
     },
   )
